Small refactor of `CollapsableContent`
Rename to `Collapsable`, to make it clearer that it includes the header as well. And rename `text` property to `title` to make it clearer that it's not a string but a react node
