+++ src/components/ThreePartHeader/ThreePartHeader.tsx
@@ -1,17 +1,17 @@
 import { Group } from '@mantine/core';
 
 interface Props {
-  text: React.ReactNode;
+  title: React.ReactNode;
   leftSection?: React.ReactNode;
   rightSection?: React.ReactNode;
 }
 
-export function ThreePartHeader({ text, leftSection, rightSection }: Props) {
+export function ThreePartHeader({ title, leftSection, rightSection }: Props) {
   return (
     <Group justify={'space-between'} gap={'xs'} wrap={'nowrap'} w={'100%'}>
       <Group gap={'xs'} flex={1} wrap={'nowrap'}>
         {leftSection && leftSection}
-        {text}
+        {title}
       </Group>
       {rightSection && rightSection}
     </Group>
