+++ src/panels/ExoplanetsPanel/ExoplanetsPanel.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { Container, Divider, Flex, Loader, ScrollArea, Text, Title } from '@mantine/core';
 
 import { useGetStringPropertyValue, useOpenSpaceApi } from '@/api/hooks';
-import { CollapsableContent } from '@/components/Collapse/CollapsableContent/CollapsableContent';
+import { Collapsable } from '@/components/Collapsable/Collapsable';
 import { FilterList } from '@/components/FilterList/FilterList';
 import { Property } from '@/components/Property/Property';
 import { PropertyOwner } from '@/components/PropertyOwner/PropertyOwner';
@@ -118,11 +118,11 @@ export function ExoplanetsPanel() {
         </Flex>
       )}
       <Divider my={'xs'} />
-      <CollapsableContent title={'Settings'}>
+      <Collapsable title={'Settings'}>
         <Property uri={HabitableZonePropertyKey} />
         <Property uri={UncertaintyDiscPropertyKey} />
         <Property uri={Size1AuRingPropertyKey} />
-      </CollapsableContent>
+      </Collapsable>
       <Divider my={'xs'}></Divider>
       <Title order={3}>Added Systems</Title>
       <ScrollArea my={'md'}>
